Binary serial communication for the arduino uno
So now binary serial communication has been implemented for the second arduino for quicker communication. The timeout allowed for the second arduino has been lowered to 5 ms, this is most likely the cause for the serial reader being so slow since it allowed the serial reader to wait until the second arduino has sent a message, which happened every 20 ms which is a big issue if the we want the serial reader to loop every 20 us.