New Arduino based controlBox code
Read the buttons attached to the pins to be defined (INPUT_PULLUP) and send the info in one byte to the portenta (controlBoxTX). The portenta can receive it with controlBoxRX. The last bit of the byte is read as 2 if there are no bytes to be read. The portenta should execute its part at less than 10 Hz (can be changed).