controlBoxTX set to the Arduino Nano R4 (Serial1) and tested
The Arduino Nano R4 does not send any characters via Serial1 apart from the programmed ones, so the boot text does not appear and spoil the communication.
The RX part has been tested again with an Arduino UNO and works as expected. The Portenta Serial port should be chosen as the connections disctate.